W1 Snacktime Minigame Logger
A partial macro, simple image comparison script that takes about 450 photos over ~30 seconds and compares those photos to photos of airborne food and food that's landed on the ground (with some redundancy); there are four positions that can be thrown to and 8 types of food. The time to win condition for that seed can be interpreted from the log based on known food combination's weights and heights.